API documentation

Constructing and accessing basic structures

The LexiconArticle

A citable object

The LexiconArticle is a citable object and therefore also implements the functions for URN citation, URN comparison, and delimited-text serialization from the CitableBase package.

SimpleLexica.LexiconArticle — Type

A simple lexicon article.

SimpleLexica.lemma — Function

Lemma or heading for a LexiconArticle.

lemma(lexart)

SimpleLexica.article — Function

Body of article for a LexiconArticle.

article(lexart)

The Lexicon

A citable collection

The Lexicon is a citable collection and therefore also implements the functions for URN filtering and delimited-text serialization from the CitableBase package, in addition to iteration from Julia base and tabular behavior from Tables.jl.

SimpleLexica.Lexicon — Type

A lexicon comprising a list of simple LexiconArticles.

SimpleLexica.lexicon — Function

Extract a Lexicon from a local file with one of Christopher Blackwell's libraries in CEX format.

lexicon(f; delimiter)

Construct a Lexicon from a list of delimited-text lines.

lexicon(linelist; delimiter)

SimpleLexica.lewis_short — Function

Construct Lewis-Short by downloading CEX from Christopher Blackwell's github repository.

lewis_short()

SimpleLexica.lsj — Function

Construct LSJ lexicon by downloading CEX from Christopher Blackwell's github repository.

lsj()

SimpleLexica.urns — Function

Extract URNs from a lexicon.

urns(lex)

SimpleLexica.lemmata — Function

Extract lemmata from a lexicon.

lemmata(lex)

SimpleLexica.articles — Function

Extract articles from a lexicon.

articles(lex)

Using a lexicon

SimpleLexica.lookup — Function

Lookup article by Cite2Urn using URN containment. The URN should uniquely identify an article in this lexicon. If it does not, the function returns nothing.

lookup(u, lexicon)

SimpleLexica.search — Function

Search lexicon lex for string s.

search(lex, s; searchscope, simplified, case_sensitive)

Optional parameters:

searchscope: one of SearchScope.LEMMA, SearchScope.ARTICLE, or SearchScope.ALL. Default: SearchScope.ALL.

simplified: a Lexicon formatted for searching. Default: nothing.

case_sensitive: true if case should be taken into account in matching. Default: true.

SimpleLexica.simplify — Function

Simplify a string for searching. Removes all non-alphabetic characters, and strips diacritics. If case_sensitive is false, makes all text lower case.

simplify(s; case_sensitive)

Simplify the lemmata and article bodies of lexicon entries for searching. Removes all non-alphabetic characters, and strips diacritics. If case_sensitive is false, makes all text lower case.

simplify(lex; case_sensitive)